[ticket/15311] Move get_temp_dir to another service
PHPBB3-15311
