+++ filesystem/filesystem.php
@@ -43,24 +43,13 @@ class filesystem implements filesystem_interface
 	protected $symfony_filesystem;
 
 	/**
-	* @var string
-	*/
-	protected $cache_temp_dir;
-
-	/**
-	* @var string
-	*/
-	protected $temp_dir;
-
-	/**
 	 * Constructor
 	 */
-	public function __construct($cache_temp_dir)
+	public function __construct()
 	{
 		$this->chmod_info			= array();
 		$this->symfony_filesystem	= new \Symfony\Component\Filesystem\Filesystem();
 		$this->working_directory	= null;
-		$this->cache_temp_dir		= $cache_temp_dir;
 	}
 
 	/**
@@ -752,34 +741,5 @@ class filesystem implements filesystem_interface
 	protected function resolve_path($path, $prefix = '', $absolute = false, $return_array = false)
 	{
 		return helper::resolve_path($path, $prefix, $absolute, $return_array);
-	}
-
-	/**
-	 * Get a temporary directory to write files
-	 *
-	 * @return string	returns the directory
-	 */
-	public function get_temp_dir()
-	{
-		if (!isset($this->temp_dir))
-		{
-			$tmp_dir = (function_exists('sys_get_temp_dir')) ? sys_get_temp_dir() : '';
-
-			// Prevent trying to write to system temp dir in case of open_basedir
-			// restrictions being in effect
-			if (empty($tmp_dir) || !@file_exists($tmp_dir) || !@is_writable($tmp_dir))
-			{
-				$tmp_dir = $this->cache_temp_dir;
-
-				if (!is_dir($tmp_dir))
-				{
-					$this->mkdir($tmp_dir, 0777);
-				}
-			}
-
-			$this->temp_dir = helper::realpath($tmp_dir);
-		}
-
-		return $this->temp_dir;
 	}
 }
